requirements.txt
numpy==1.14.4
Pillow==5.1.0
protobuf==3.5.2.post1
six==1.11.0
tensorboardX==1.2
torch==0.4.1
torchvision==0.2.1
tqdm==4.23.4
scikit-image
